# PDS-master/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "UniRooms API",
    "description": "Documentation for the UniRooms API",
    "version": "1.0.0"
  },
  "host": "unirooms-preview.herokuapp.com",
  "basePath": "/",
  "schemes": ["https"],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "x-access-token",
      "description": "Authentication token"
    }
  },
  "paths": {
...
    "/api/auth/passwordreset/set": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "token": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}